fix: 장바구니 삭제 버그 및 전체 선택 수정
